+++ sample/src/main/java/floatingmuseum/sample/sonic/App.java
@@ -45,6 +45,7 @@ public class App extends Application {
                 .setConnectTimeout(3000)
                 .setDirPath(getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS).getAbsolutePath())
                 .setLogEnabled()
+                .setBroadcastAction("FloatingMuseum")
                 .init(this);
     }
 }